- fabfile update

+++ fabfile.py
@@ -29,12 +29,13 @@ def push():
 def pull(env_name):
     with cd('/opt/shoutit_api_{}/api'.format(env_name)):
         run('git pull')
-        run('/opt/shoutit_api_{}/bin/pip install    -r src/requirements/common_noupdate.txt'.format(env_name))
-        run('/opt/shoutit_api_{0}/bin/pip install -U -r src/requirements/{0}.txt'.format(env_name))
-        # run('/opt/shoutit_api_{}/bin/python src/manage.py test'.format(env))
-        run('/opt/shoutit_api_{}/bin/python src/manage.py migrate'.format(env_name))
-        if confirm("Clear all logs?"):
-            run("find /opt/shoutit_api_{}/log/. -type f -exec cp /dev/null {{}} \;".format(env_name))
+        if not confirm('Fast deploy [pull and restart only]?'):
+            run('/opt/shoutit_api_{}/bin/pip install    -r src/requirements/common_noupdate.txt'.format(env_name))
+            run('/opt/shoutit_api_{0}/bin/pip install -U -r src/requirements/{0}.txt'.format(env_name))
+            # run('/opt/shoutit_api_{}/bin/python src/manage.py test'.format(env))
+            run('/opt/shoutit_api_{}/bin/python src/manage.py migrate'.format(env_name))
+            if confirm("Clear all logs?"):
+                run("find /opt/shoutit_api_{}/log/. -type f -exec cp /dev/null {{}} \;".format(env_name))
         run('supervisorctl restart all')
